+++ NEWS.md
@@ -1,3 +1,5 @@
+# cppdoubles (development version)
+
 # cppdoubles 0.3.0
 
 ### Breaking changes